adjustment table and add some component

+++ routes/web.php
@@ -14,23 +14,23 @@ Route::middleware(['auth', 'verified'])->group(function () {
 
 
     Route::prefix('users')->name('users.')->group(function () {
-        Route::get('/', [UserController::class, 'index'])->name('index');
-        Route::get('/create', [UserController::class, 'create'])->name('create');
-        Route::post('/', [UserController::class, 'store'])->name('store');
-        Route::get('/{user}', [UserController::class, 'show'])->name('show');
-        Route::get('/{user}/edit', [UserController::class, 'edit'])->name('edit');
-        Route::put('/{user}', [UserController::class, 'update'])->name('update');
-        Route::delete('/{user}', [UserController::class, 'destroy'])->name('destroy');
+        Route::get('/', [UserController::class, 'index'])->middleware('permission:users.view')->name('index');
+        Route::get('/create', [UserController::class, 'create'])->middleware('permission:users.create')->name('create');
+        Route::post('/', [UserController::class, 'store'])->middleware('permission:users.create')->name('store');
+        Route::get('/{user}', [UserController::class, 'show'])->middleware('permission:users.show')->name('show');
+        Route::get('/{user}/edit', [UserController::class, 'edit'])->middleware('permission:users.edit')->name('edit');
+        Route::put('/{user}', [UserController::class, 'update'])->middleware('permission:users.edit')->name('update');
+        Route::delete('/{user}', [UserController::class, 'destroy'])->middleware('permission:users.delete')->name('destroy');
     });
 
     Route::prefix('roles')->name('roles.')->group(function () {
-        Route::get('/', [RoleController::class, 'index'])->name('index');
-        Route::get('/create', [RoleController::class, 'create'])->name('create');
-        Route::post('/', [RoleController::class, 'store'])->name('store');
-        Route::get('/{user}', [RoleController::class, 'show'])->name('show');
-        Route::get('/{user}/edit', [RoleController::class, 'edit'])->name('edit');
-        Route::put('/{user}', [RoleController::class, 'update'])->name('update');
-        Route::delete('/{user}', [RoleController::class, 'destroy'])->name('destroy');
+        Route::get('/', [RoleController::class, 'index'])->middleware('permission:roles.view')->name('index');
+        Route::get('/create', [RoleController::class, 'create'])->middleware('permission:roles.create')->name('create');
+        Route::post('/', [RoleController::class, 'store'])->middleware('permission:roles.create')->name('store');
+        Route::get('/{user}', [RoleController::class, 'show'])->middleware('permission:roles.show')->name('show');
+        Route::get('/{user}/edit', [RoleController::class, 'edit'])->middleware('permission:roles.edit')->name('edit');
+        Route::put('/{user}', [RoleController::class, 'update'])->middleware('permission:roles.edit')->name('update');
+        Route::delete('/{user}', [RoleController::class, 'destroy'])->middleware('permission:roles.delete')->name('destroy');
     });
 });
